Refactor RawDataView into a new component
